refactor: 디렉토리 구조 최종 완료(페이지별)
