+++ frontend/src/components/global/Header.vue
@@ -26,7 +26,7 @@
                 Community
                 </button>
                 <ul class="dropdown-menu menu-item" aria-labelledby="dropdownMenuButton1">
-                <li><a class="dropdown-item" href="/home">습관 공유 커뮤니티</a></li>
+                <li><a class="dropdown-item" href="/routine-community">습관 공유 커뮤니티</a></li>
                 <li><a class="dropdown-item" href="/post-community">인증 커뮤니티</a></li>
                 </ul>
             </div>
